refactor: share the input instance and style state types between platforms
