+++ packages/react-native-enriched-markdown/src/EnrichedMarkdownTextInput.tsx
@@ -44,6 +44,12 @@ import { normalizeMenuItem } from './normalizeMenuItem';
 import { toNativeRegexConfig } from './utils/regexParser';
 import { TextInputState } from './utils/textInputState';
 import type { RefObject } from 'react';
+import type {
+  CaretRect,
+  HeadingLevel,
+  MarkdownTextInputInstance,
+  StyleState,
+} from './types/MarkdownTextInputInstance';
 
 type NativeRef = HostInstance;
 
@@ -53,7 +59,9 @@ export interface LinkStyle {
   backgroundColor?: string;
 }
 
-export type HeadingLevel = 1 | 2 | 3 | 4 | 5 | 6;
+export type { HeadingLevel, StyleState, CaretRect };
+export type EnrichedMarkdownTextInputInstance =
+  MarkdownTextInputInstance<HostInstance>;
 
 const VALID_HEADING_LEVELS = new Set<number>([1, 2, 3, 4, 5, 6]);
 
@@ -102,18 +110,6 @@ export interface MarkdownTextInputStyle {
   };
 }
 
-export interface StyleState {
-  bold: { isActive: boolean };
-  italic: { isActive: boolean };
-  underline: { isActive: boolean };
-  strikethrough: { isActive: boolean };
-  spoiler: { isActive: boolean };
-  link: { isActive: boolean };
-  heading: { isActive: boolean; level: HeadingLevel };
-  unorderedList: { isActive: boolean; depth: number };
-  orderedList: { isActive: boolean; depth: number };
-}
-
 export interface ContextMenuItem {
   text: string;
   onPress: (event: {
@@ -123,42 +119,6 @@ export interface ContextMenuItem {
   }) => void;
   icon?: string;
   visible?: boolean;
-}
-
-export interface CaretRect {
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-}
-
-export interface EnrichedMarkdownTextInputInstance {
-  focus: () => void;
-  blur: () => void;
-  measure: HostInstance['measure'];
-  measureInWindow: HostInstance['measureInWindow'];
-  measureLayout: HostInstance['measureLayout'];
-  setValue: (markdown: string) => void;
-  setSelection: (start: number, end: number) => void;
-  toggleBold: () => void;
-  toggleItalic: () => void;
-  toggleUnderline: () => void;
-  toggleStrikethrough: () => void;
-  toggleSpoiler: () => void;
-  toggleHeading: (level: HeadingLevel) => void;
-  toggleUnorderedList: () => void;
-  toggleOrderedList: () => void;
-  indentList: () => void;
-  outdentList: () => void;
-  setLink: (url: string) => void;
-  insertLink: (text: string, url: string) => void;
-  insertText: (text: string) => void;
-  insertMention: (displayText: string, url: string) => void;
-  startMention: (indicator: string) => void;
-  removeLink: () => void;
-  copyToClipboard: () => void;
-  getMarkdown: () => Promise<string>;
-  getCaretRect: () => Promise<CaretRect>;
 }
 
 /**
